objection_detection_from_picture
This is the code I wrote for object detection in an image using OpenCV. It can locate and draw a rectangle around any object in a photo, like this wheel on the NASA space technology. It was a great way to practice my coding skills and learn more about OpenCV and its methods. See the code and the result in the following slides.